Message editor and user message list

+++ src/main/java/com/example/twitter/repos/MessagesRepo.java
@@ -8,6 +8,6 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
 
-public interface MessagesRepo extends CrudRepository<Message, Integer> {
+public interface MessagesRepo extends CrudRepository<Message, Long> {
         List<Message> findByTag(String tag);
 }
